requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
requests==2.27.1
urllib3==1.26.9
